+++ spotinst_sdk2/test/elastigroup/aws/test_aws_eg_model.py
@@ -402,7 +402,7 @@ class AwsElastigroupTestBlockDeviceMapping(AwsElastigroupTestCase):
 
 class AwsElastigroupTestEbs(AwsElastigroupTestCase):
     def runTest(self):
-        ebs = [EBS(snapshot_id="snp-1")]
+        ebs = [EBS(snapshot_id="snp-1", throughput=500)]
         block_device_mappings = [BlockDeviceMapping(ebs=ebs)]
         launch_specification = LaunchSpecification(block_device_mappings=block_device_mappings)
         compute = Compute(launch_specification=launch_specification)
@@ -414,7 +414,8 @@ class AwsElastigroupTestEbs(AwsElastigroupTestCase):
 
         actual_request_json = formatted_group_dict['group']['compute']['launchSpecification']['blockDeviceMappings'][0]['ebs'][0]
         expected_request_json = {
-            'snapshotId': "snp-1"
+            'snapshotId': "snp-1",
+            'throughput': 500
         }
 
         self.assertDictEqual(actual_request_json, expected_request_json)